gotta fix the overflow check atoi

+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: mdursun <mdursun@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/07 16:32:34 by mdursun           #+#    #+#             */
-/*   Updated: 2024/11/28 13:08:58 by mdursun          ###   ########.fr       */
+/*   Updated: 2024/12/15 21:07:39 by mdursun          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -42,7 +42,7 @@ char	*ft_itoa(int n);
 char	*ft_strtrim(const char *s1, const char *set);
 char	*ft_strjoin(char const *s1, char const *s2);
 char	*ft_substr(char const *s, unsigned int start, size_t len);
-int		ft_atoi(const char *str);
+long		ft_atoi(const char *str);
 char	*ft_strnstr(const char *haystck, const char *needle, size_t n);
 void	*ft_calloc(size_t count, size_t size);
 char	*ft_strdup(const char *s1);
